requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
PyPDF2==3.0.1
google-generativeai==0.5.2
SQLAlchemy==2.0.29
aiosqlite==0.20.0
